+++ includes/plugins/advanced-custom-fields/advanced-custom-fields.php
@@ -11,7 +11,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    1.0.0
- * @version  1.0.0
+ * @version  1.3.1
  *
  * Contents:
  *
@@ -27,7 +27,7 @@
  * 1) Requirements check
  */
 
-	if ( ! function_exists( 'register_field_group' ) || ! is_admin() ) {
+	if ( ! function_exists( 'acf_add_local_field_group' ) || ! is_admin() ) {
 		return;
 	}